refactor(Sprint10): `field-sizing` 속성 대신 <span>을 활용해서 <Input> text 가운데 정렬
`field-sizing`은 비표준 속성으로 지원하지 않는 브라우저가 있으므로 사용하지 않습니다.
