+++ my-app/components/todo/todo-detail-header.module.css
@@ -1,4 +1,4 @@
-.todoDetailTitle {
+.todoDetailHeader {
   background-color: white;
   border: 2px solid var(--color-state-900);
   border-radius: 24px;
@@ -8,21 +8,31 @@
   align-items: center;
   gap: 16px;
   height: 64px;
-  text-decoration: underline;
 }
-.todoDetailTitle.checked {
+.todoDetailHeader.checked {
   background-color: var(--color-violet-200);
 }
 
-.todoDetailTitle input {
-  field-sizing: content;
+.todoTitle,
+.todoTitleMirror {
+  font-size: 20px;
+  font-weight: 800;
+  line-height: 100%;
+  text-decoration: underline;
+}
+
+.todoTitle {
   padding: 0;
   background: none;
   border: none;
   outline: none;
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 100%;
+}
+
+.todoTitleMirror {
+  position: absolute;
+  visibility: hidden;
+  white-space: pre;
+  pointer-events: none;
 }
 
 .checkImage {
